Move Mission component to root directoyr
